Working version of the mouse parser.
This takes a CSV file of (time_us, x, y) and converts this into paths that the mouse followed for replay. The time between recordings is expected to be consistent throughout the entire input file, not checked on a per path basis. Output creates a new file, doesn't append to some existing map.