Refactor mouse crate to allow multiple binaries.
It seems that placing a main in src/bin/ automatically makes a binary of that files name, and there can be multiple of these. https://stackoverflow.com/questions/36604010/how-can-i-build-multiple-binaries-with-cargo

Also once again the hardest part was figuring out the imports. Seems like we need to place all of the files I write into lib.rs for 'pub mod'