Improve replay.
Now MouseMover can stitch multiple paths together to get progressively closer to the goal. This is limited to avoid an infinite loop if there is no valid path found.

To avoid stuttering around the point, we can move a single delta's worth explicitly. I estimate this to be about 10 pixels which should be safe. Being able to place pixel perfect will help make other places in the code simpler since we know exactly where the mouse will go.

Overall we cannot always click on the exact same pixel, but that randomness can come from elsewhere, not as part of the MouseMover, or perhaps that will be an explicit add on.