Begin work on class to wrap screenshotting.
I now block on capturing the frame forever, but for now I'm only planning to run on my PC so let's assume I have a screen :)

Successfully able to crop an image. The vast majority of time running this was for saving the image, so need to find a faster method of getting output for rapid testing.