Create wrapper classes for using scrap.
Each one has some interesting ownership issues Capturer.flip is a straightforward issue where the compiler need to be updated to understand a lifetime in a loop. Frame is an interesting situation since sometimes it will own the data, but sometimes not (this is to avoid uneccessarily copying in the entire Frame.

I am moving {,Delta}Position to be shared types across crates in the workspace so this change will probably break mouse/