Made significant progress on screen handling.
In light of not being able to use the polonius flag we have to make some changes to the library, but they aren't too bad. Since the current design is to randomly look for matching pixels I can probably get away with a single function interface for the screen crate for the actual bot.

Still think all the fancy stuff I did with Frame/Capturer were worth it to learn about rust (which is the point of this project) :)