Update min tolerance and cheat distance.
Now start looking for paths with a tolerance of 1. We expand exponentially and search seems to be pretty fast, so it makes sense to start with a tight tolerance so the path we find will get us as close as possible without cheating.

MAX_CHEAT_DISTANCE - move back to 20. With the earlier fix preventing stutter at the end this is looking ok to me.