New version of close_chatbox.
This one checks if the box is open by looking for the corners. It's a bit finicky though since the colors on the corners change so quickly over just a few pixels. So the screen being highlighted or not makes a big difference. The reason for this change overall was to avoid excess moving an clicking since just analyzing a screenshot is relatively fast.