Rafactor for checking pixels.
Moved to FuzzyPixel automatically implementing a match that includes the ratios of the channels. Previously I had begun to implement separate checkers for white and blue for the action letters. This should hopefully allow us to expand the range of values per chanel, while better enforcing the actual color we want to find.