Refine checking for letter matches.
Partly this was an improvement in FizzyPixel matching, since the way I was calculating min/max ratios was insufficient. Another notable change is that now spaces can be (4, 5, 6) pixels to account for the fact that they actually do seem to change (or at least something does). I have seen images where the letters extend beyond the expectations and where the opposite is true. Hopefully I didn't overfit for this specific word.