Chopping wood is working moderately well.
I am working on rotating the screen, which it turned out was a story. The button inputbot fakes corresponds to the number bad arrow, which (even when num lk is off) doesn't seem to rotate the screen including when I manually tested. Luckily runelite has a plugin so I can rotate with WASD, which requires continually sending the press signal, not pressing and holding. Thanks to 'sudo showkey -s' for helping figure this all out