Update mouse crate.
inputbot crate is good for mouse handling, but uinput crate is better for handling keyboard inputs. I think at this point it may be a bit of a misnomer to call the crate mouse, and I may want to work on renaming the whole thing to input or controller or something (inputbot and uinput are both taken tho). This will also involve a move to have the InputBot struct be the only thing exported from the (current) mouse crate