chopping oak is working pretty well.
Working on creating Player class. Having some issues due to mutable borrows, since now capturer and inputbot are fields, so accessing both creates a double mutable borrow. I need to decide if to use RefCell on the Player (wrap capturer and inputbot) or to use it within each of those clases.