Add minimap movement as well as worldmap actions.
I was able to successfully travel to the bank when it was in the minimap. Next is adding motion following the worldmap. Polar coordinates are really cool, allowed me to cover most of the minimap well instead of needing to either subscribe or circumscribe the minimap with a bounding box.