Work on bank locations.
This was the most complicated yet. Seems to be pretty good for both fullscreen and 'minimized'. I think that my plan for the bank isn't to require pixel perfect placement though, since I won't analyze items in the bank. Rather I will have the user withdraw based on the location in the bank, that way we can afford to be less accurate. This requires the user to get the slot_index of the items in the bank they wish to withdraw. Also pressing the Quantity button doesn't need to be pixel perfect so I think I am good enough here.