Switch the waiting from DescribeAction to check the frame.
This is because some tasks take variable amounts of time and may not happen. Therefore we check that a condition has actually been met before completing. We should probably add a config to Player to have a max await time to panic if we fail on the await for over X time.