Begin transition from DescribeAction to Action.
This new interface will allow much more control since now we are passing all of the facets of the player instead of just frame/framehandler. We break them apart, as opposed to passing Player, since inputbot and capturer are both mutable, but should be independent of each other.

Consider merging capturer and framehandler in screenhandler so that each crate will have a singular interface...