Update mouse movements.
Remove condition from parser not to have paths of length 0. Add this to controller. This is because the parser just parses paths and doesn't need to decide which lengths are good. Also the parser loop is already fairly complicated so I'd rather not put this there.

Also don't start mouse paths at tolerance 1, rather CHEAT_DIST-1. This means we will always have a few options to pick the fastest one. The mouse is now moving very fast.