Switch gitignore to allow mouse tracking info.
This file isn't huge and it doesn't update so much so I am no longer worried about it clogging up the git history. These files are needed to run the bot so it makes setup much easier just to have them handy.

We still exclude the screenshots since those update more often and are not needed to run the bot so not sure if they would add much.